08-10-2025 Fixing Time of Alignment
